retry_backoff: Extract get_retry_backoff_seconds to a leaf module.
base_page_params.ts needs the same backoff logic but cannot import
util.ts directly: util imports i18n, and i18n reads page_params at
module top level, so the cycle would leave i18n seeing an
undefined page_params at init time.

Move the function to a leaf module that imports only zod, then
re-export it from util.ts to preserve the public API and the
let/rewire test seam.
